# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: |
    REST API для управления банковскими картами с поддержкой:
    - Аутентификации и авторизации (JWT)
    - Управления картами (создание, блокировка, активация)
    - Переводов между картами
    - Управления пользователями
  version: 1.0.0
  contact:
    name: API Support
    email: support@bankcards.example.com
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:8080
    description: Local Docker environment
...
security:
  - bearerAuth: []
paths:
...
  /api/users:
    get:
      tags:
        - Пользователи (ADMIN)
      summary: Получить всех пользователей
      description: Получает список всех пользователей с пагинацией (только ADMIN)
      operationId: getAllUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserResponseDto'
  /api/users/{id}/role:
    patch:
      tags:
        - Пользователи (ADMIN)
      summary: Назначить роль пользователю
      description: Изменяет роль пользователя (только ADMIN)
      operationId: updateUserRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdateDto'
            example:
              role: "ROLE_ADMIN"
      responses:
        '200':
          description: Роль успешно изменена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '404':
          description: Пользователь не найден
  /api/users/{id}:
    delete:
      tags:
        - Пользователи (ADMIN)
      summary: Удалить пользователя
      description: Удаляет пользователя из системы (только ADMIN)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь успешно удален
        '404':
          description: Пользователь не найден
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации. Получите токен через /auth/sign-in или /auth/sign-up
  schemas:
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: "john_doe"
        email:
          type: string
          format: email
          description: Email пользователя
          example: "john@example.com"
        role:
          type: string
          enum: [ROLE_USER, ROLE_ADMIN]
          description: Роль пользователя
          example: "ROLE_USER"
    UserRoleUpdateDto:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [ROLE_USER, ROLE_ADMIN]
          description: Новая роль пользователя
          example: "ROLE_ADMIN"
...
    PageUserResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseDto'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalPages:
          type: integer
        totalElements:
          type: integer
        last:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        sort:
          $ref: '#/components/schemas/Sort'
        offset:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...